:recycle: (#123) Move sidebar components to sidebar folder
