+++ frontend/src/App.svelte
@@ -5,11 +5,9 @@
     accessToken,
     LoginButton,
   } from "@dopry/svelte-oidc";
-  import Navbar from "./components/navbar/Navbar.svelte";
-  import GroupLogos from "./components/groups/GroupLogos.svelte";
   import Footer from "./components/footer/Footer.svelte";
   import PunishmentGrid from "./components/punishments/PunishmentGrid.svelte";
-  import Sidebar from "./components/filters/Sidebar.svelte";
+  import Sidebar from "./components/sidebar/Sidebar.svelte";
   import { getOnlineProfile, getMyOnlineGroups } from "./lib/api";
 </script>
 
@@ -32,14 +30,7 @@
     redirect_uri="http://localhost:3000"
     post_logout_redirect_uri="http://localhost:3000"
     scope="openid profile onlineweb4"
-    extraOptions="{{
-      metadataUrl:
-        'https://old.online.ntnu.no/openid/.well-known/openid-configuration',
-      filterProtocolClaims: true,
-      loadUserInfo: true,
-      silent_redirect_uri: 'http://localhost:3000',
-      revokeAccessTokenOnSignout: true,
-    }}"
+    extraOptions="{{ metadataUrl: 'https://old.online.ntnu.no/openid/.well-known/openid-configuration', filterProtocolClaims: true, loadUserInfo: true, silent_redirect_uri: 'http://localhost:3000', revokeAccessTokenOnSignout: true }}"
   >
     {#if $isAuthenticated}
       <div class="body_content">
@@ -56,12 +47,16 @@
                 >
                   <img src="assets/icons/sad.png" alt="sadface" width="100" />
                   <p class="ml-4">
-                    Kunne ikke finne noen grupper du er medlem i. <br /> Ta
-                    kontakt med
+                    Kunne ikke finne noen grupper du er medlem i.
+                    <br />
+                    Ta kontakt med
                     <a
                       class="hover:underline"
-                      href="mailto:dotkom@online.ntnu.no">Dotkom</a
-                    > om dette er feil.
+                      href="mailto:dotkom@online.ntnu.no"
+                    >
+                      Dotkom
+                    </a>
+                    om dette er feil.
                   </p>
                 </div>
               {/if}
