+++ frontend/src/components/sidebar/GroupPicker.svelte
