+++ frontend/src/components/sidebar/Logout.svelte
