+++ frontend/src/components/sidebar/SearchUser.svelte
